# src/main/resources/swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"Minimerce is an E-Commerce API Platform.<br/>Provides product registration and ordering, state management, and external API creation.",
    "version":"0.0.1",
    "title":"Swagger Minimerce",
    "contact":{
      "email":"geminikims@gmail.com"
    },
    "license":{
      "name":"Apache 2.0",
      "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host":"localhost:8080",
  "basePath":"/v1",
  "tags":[
    {
      "name":"order-controller",
      "description":"Order Controller"
    },
    {
      "name":"customer-controller",
      "description":"Customer Controller"
    }],
  "paths":{
    "/customers/{customerId}/orders":{
      "get":{
        "tags":[
          "customer-controller"],
        "summary":"findCustomerOrders",
        "operationId":"findCustomerOrdersUsingGET",
        "consumes":[
          "application/json;charset=UTF-8"],
        "produces":[
          "*/*"],
        "parameters":[
          {
            "name":"customerId",
            "in":"path",
            "description":"customerId",
            "required":true,
            "type":"integer",
            "format":"int64"
          }],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/ApiResponse"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      }
    },
    "/orders":{
      "get":{
        "tags":[
          "order-controller"],
        "summary":"findOrder",
        "operationId":"findOrderUsingGET",
        "consumes":[
          "application/json;charset=UTF-8"],
        "produces":[
          "application/json;charset=UTF-8"],
        "parameters":[
          {
            "name":"orderId",
            "in":"query",
            "description":"orderId",
            "required":false,
            "type":"integer",
            "format":"int64"
          },
          {
            "name":"clientOrderId",
            "in":"query",
            "description":"clientOrderId",
            "required":false,
            "type":"integer",
            "format":"int64"
          }],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/ApiResponse"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "post":{
        "tags":[
          "order-controller"],
        "summary":"order",
        "operationId":"orderUsingPOST",
        "consumes":[
          "application/json;charset=UTF-8"],
        "produces":[
          "application/json;charset=UTF-8"],
        "parameters":[
          {
            "in":"body",
            "name":"request",
            "description":"request",
            "required":true,
            "schema":{
              "$ref":"#/definitions/OrderRequest"
            }
          }],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/ApiResponse"
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      }
    }
  },
  "securityDefinitions":{
    "api-key":{
      "type":"apiKey",
      "name":"minimerce-api-key",
      "in":"header"
    }
  },
  "definitions":{
    "ApiResponse":{
      "type":"object",
      "properties":{
        "data":{
          "type":"object"
        },
        "result":{
          "$ref":"#/definitions/ApiResult"
        }
      }
    },
    "ApiResult":{
      "type":"object",
      "properties":{
        "code":{
          "type":"integer",
          "format":"int32"
        },
        "isError":{
          "type":"boolean"
        },
        "message":{
          "type":"string"
        },
        "status":{
          "type":"integer",
          "format":"int32"
        }
      }
    },
    "OrderRequest":{
      "type":"object",
      "properties":{
        "clientOrderId":{
          "type":"integer",
          "format":"int64"
        },
        "customerId":{
          "type":"integer",
          "format":"int64"
        },
        "details":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/OrderRequestDetail"
          }
        },
        "orderedAt":{
          "type":"string",
          "format":"date-time"
        },
        "price":{
          "type":"integer",
          "format":"int32"
        }
      }
    },
    "OrderRequestDetail":{
      "type":"object",
      "properties":{
        "dealId":{
          "type":"integer",
          "format":"int64"
        },
        "optionId":{
          "type":"integer",
          "format":"int64"
        },
        "price":{
          "type":"integer",
          "format":"int32"
        },
        "quantity":{
          "type":"integer",
          "format":"int32"
        },
        "unitPrice":{
          "type":"integer",
          "format":"int32"
        }
      }
    }
  }
}